+++ gradle.properties
@@ -5,6 +5,6 @@ exposedVersion=0.55.0
 mysqlVersion=8.0.33
 hikaricpVersion=5.0.1
 
-
+logbackClassicVersion=1.5.11
 testCoroutinesVersion=1.9.0
 junitVersion=5.11.2
