+++ playground/src/main/kotlin/Main.kt
@@ -2,11 +2,8 @@ package po.playground
 
 
 import io.github.cdimascio.dotenv.dotenv
-import po.db.data_service.models.ConnectionModel
-import po.playground.projects.data_service.startDataService
 
 
-import po.api.rest_service.server.ApiServer
 import po.playground.projects.rest_service.startApiServer
