+++ rest_service/src/main/kotlin/po/api/rest_service/plugins/Logging.kt
@@ -2,7 +2,7 @@ package po.api.rest_service.plugins
 
 import io.ktor.server.application.createApplicationPlugin
 import po.api.rest_service.logger.LoggingService
-import po.api.rest_service.server.ApiServer
+import po.api.rest_service.RestServer
 
 val LoggingPlugin = createApplicationPlugin(name = "LoggingPlugin") {
     val apiLogger = LoggingService()
@@ -18,6 +18,6 @@ val LoggingPlugin = createApplicationPlugin(name = "LoggingPlugin") {
     onCallRespond { call, body ->
        //Some Logic after sending data
     }
-    application.attributes.put(ApiServer.loggerKey, apiLogger)
+    application.attributes.put(RestServer.loggerKey, apiLogger)
     println("Logger registered in Application: ${application.hashCode()}")
 }
